pyproject.toml
[project]
name = "life-ai"
version = "0.1.0"
description = "Personal life observer - camera-based daily activity tracker and summarizer"
authors = [
    { name = "andy", email = "takuyarossi77@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "opencv-python-headless>=4.9",
    "click>=8.1",
    "rich>=13.0",
    "google-genai>=1",
]

[project.scripts]
life = "life.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["life"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
